I added follow-parent to resource
if follow-parent is true
then
the originator pid is the parent of what it would otherwise follow.

if follow-parent is false then then originator pid is the process that created this resource.  we can garbage collect this resource starting when that process is through.
this is useful for most scripts.

however for more interactive use we anticipate that the script might echo the resource.
we want garbage collection to wait until the parent of that script is through.